requirements.txt
flask>=3.0.0
openai>=1.30.0
python-dotenv>=1.0.1
pypdf>=4.2.0
python-docx>=1.1.2
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
numpy>=1.26.0
